@@ -11,7 +11,7 @@ except ImportError:
 import smmap
 
 if os.path.exists("README.md"):
-    long_description = open('README.md', "r", encoding="utf-8").read().replace('\r\n', '\n')
+    long_description = open('README.md', encoding="utf-8").read().replace('\r\n', '\n')
 else:
     long_description = "See https://github.com/gitpython-developers/smmap"
 
@@ -26,7 +26,7 @@ setup(
     license="BSD",
     packages=find_packages(),
     zip_safe=True,
-    python_requires=">=3.5",
+    python_requires=">=3.6",
     classifiers=[
         "Development Status :: 5 - Production/Stable",
         "Environment :: Console",
@@ -38,15 +38,13 @@ setup(
         "Operating System :: MacOS :: MacOS X",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3 :: Only",
     ],
     long_description=long_description,
     long_description_content_type='text/markdown',
-    tests_require=('nose', 'nosexcover'),
-    test_suite='nose.collector'
 )
