@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # smmap documentation build configuration file, created by
 # sphinx-quickstart on Wed Jun  8 15:14:25 2011.
@@ -38,8 +37,8 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'smmap'
-copyright = u'2011, Sebastian Thiel'
+project = 'smmap'
+copyright = '2011, Sebastian Thiel'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -173,8 +172,8 @@ htmlhelp_basename = 'smmapdoc'
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title, author, documentclass [howto/manual]).
 latex_documents = [
-    ('index', 'smmap.tex', u'smmap Documentation',
-     u'Sebastian Thiel', 'manual'),
+    ('index', 'smmap.tex', 'smmap Documentation',
+     'Sebastian Thiel', 'manual'),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
